+++ internal/domain/products/model.go
@@ -8,9 +8,9 @@ import (
 const TableNameProduct = "products"
 
 type Product struct {
-	ID             int32           `gorm:"column:id;primaryKey;autoIncrement:true"`
+	ID             int64           `gorm:"column:id;primaryKey;autoIncrement:true"`
 	Name           string          `gorm:"column:name;not null"`
-	UnitID         int32           `gorm:"column:unit_id"`
+	UnitID         int64           `gorm:"column:unit_id"`
 	WholesalePrice decimal.Decimal `gorm:"column:wholesale_price"`
 	RetailPrice    decimal.Decimal `gorm:"column:retail_price"`
