+++ internal/domain/sales_and_products/model.go
@@ -8,7 +8,7 @@ const TableNameSalesAndProduct = "sales_and_products"
 
 type SalesAndProduct struct {
 	SaleID    int64 `gorm:"column:sale_id;primaryKey"`
-	ProductID int32 `gorm:"column:product_id;primaryKey"`
+	ProductID int64 `gorm:"column:product_id;primaryKey"`
 	Amount    int32 `gorm:"column:amount;not null"`
 
 	Product products.Product `gorm:"foreignKey:ProductID;references:ID"`
