+++ internal/domain/standart_equipment/model.go
@@ -7,8 +7,8 @@ import (
 const TableNameStandartEquipment = "standart_equipment"
 
 type StandartEquipment struct {
-	GateTypeID int32 `gorm:"column:gate_type_id;primaryKey"`
-	OptionID   int32 `gorm:"column:option_id;primaryKey"`
+	GateTypeID int64 `gorm:"column:gate_type_id;primaryKey"`
+	OptionID   int64 `gorm:"column:option_id;primaryKey"`
 	Amount     int32 `gorm:"column:amount"`
 
 	Option options.Option `gorm:"foreignKey:OptionID;references:ID"`
